+++ src/main/java/blackjack/model/participant/Participant.java
@@ -1,5 +1,6 @@
 package blackjack.model.participant;
 
+import blackjack.model.FinishedState;
 import blackjack.model.card.Card;
 import blackjack.model.card.CardDeck;
 import blackjack.model.state.State;
